+++ projects/configgen/configgen/generators/dolphin/dolphinControllers.py
@@ -182,6 +182,9 @@ def generateControllerConfigAny(playersControllers: ControllerPerPlayer, filenam
             f.write("Main Stick/Dead Zone = 25.000000000000000" + "\n")
             f.write("C-Stick/Dead Zone = 25.000000000000000" + "\n")
 
+            if system.Rumble:
+                f.write("Rumble/Motor = Sine|Square|Triangle|Strong|Weak" + "\n")
+
         for inp in pad.AvailableInput:
 
             if inp.Item in anyMapping: